Upgraded to scodec-bits 1.1.0

+++ src/main/scala/scalaz/stream/io.scala
@@ -301,7 +301,7 @@ object io {
                 go(offset, length, read)
             } else {
               val chunk = chunks.head
-              val remaining = chunk.length - index
+              val remaining = chunk.length.toInt - index
 
               if (length <= remaining) {
                 chunk.copyToArray(buffer, offset, index, length)
